Refactor. Separated quiz repository into 4 repositories for each controller.
