+++ QuizyfyAPI-Tests/QuizControllerTest.cs
@@ -1,0 +1,22 @@
+using QuizyfyAPI.Controllers;
+using System;
+using Xunit;
+
+namespace QuizyfyAPI_Tests
+{
+    public class QuizControllerTest
+    {
+        QuizzesController _quizzesController;
+
+        public QuizControllerTest()
+        {
+            _quizzesController = new QuizzesController();
+        }
+
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
