+++ QuizyfyAPI.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29009.5
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuizyfyAPI", "QuizyfyAPI\QuizyfyAPI.csproj", "{3490FA6F-8118-4A5F-924F-4DC95740399E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuizyfyAPI", "QuizyfyAPI\QuizyfyAPI.csproj", "{3490FA6F-8118-4A5F-924F-4DC95740399E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuizyfyAPI-Tests", "QuizyfyAPI-Tests\QuizyfyAPI-Tests.csproj", "{0D305260-531D-4259-BC1F-C474DF6A0DB2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{3490FA6F-8118-4A5F-924F-4DC95740399E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3490FA6F-8118-4A5F-924F-4DC95740399E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3490FA6F-8118-4A5F-924F-4DC95740399E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0D305260-531D-4259-BC1F-C474DF6A0DB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D305260-531D-4259-BC1F-C474DF6A0DB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D305260-531D-4259-BC1F-C474DF6A0DB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D305260-531D-4259-BC1F-C474DF6A0DB2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
