+++ QuizyfyAPI/Data/Repositories/IRepository.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace QuizyfyAPI.Data
+{
+    public interface IRepository
+    {
+        void Add<T>(T entity) where T : class;
+        void Delete<T>(T entity) where T : class;
+        Task<bool> SaveChangesAsync();
+    }
+}
+
